Add WiFi Wardriver, remove WiFi Sniffer --nobuild Wardriver is an updated version of sniffer. The name is just better lol
