+++ applications/external/wardriver/helpers/minmea.c
