+++ applications/external/wardriver/helpers/minmea.h
