+++ applications/external/wardriver/icons/up.png
