--- applications/external/wifisniffer/application.fam
@@ -1,15 +1,0 @@
-App(
-    appid="wifisniffer",
-    name="[ESP32GPS] Wifi Sniff",
-    apptype=FlipperAppType.EXTERNAL,
-    entry_point="wifisniffer_app",
-    stack_size=2 * 1024,
-    fap_icon="sniff.png",
-    fap_category="WiFi",
-    fap_author="@Sil333033",
-    fap_weburl="https://github.com/Flipper-XFW/Xtreme-Apps/tree/dev/wifisniffer",
-    fap_version="1.0",
-    fap_description="Sniff WiFi access points with GPS location coordinates",
-    fap_icon_assets="icons",
-    fap_icon_assets_symbol="wifisniffer",
-)
